matrix: Messure animation lenght in tics
This is less precise in terms of how long the whole animation will
take, because on slower machines, the drawing takes more time which
is not accounted into duration. However, it makes sure, everyone
sees the whole animation even in case the program lags for a moment
(so the number of steps will be constant).

Related to: https://github.com/KanoComputing/peldins/issues/783