Boot: Prevent infinite loop when adding users
When running `kano-init boot`, it checks the status file and loops
through the phases until the status is set to disabled. This function
only uses the in-memory instance of the status, which means that if the
status is set from another process, as happens during the onboarding,
then the status instance is never updated and hence says in a looping
state. Prevent this from occurring by limiting the number of loops that
can be done before giving up.

Ideally we would remove the need to loop at all, but this patch is
unobtrusive.