Remove reverse dependency of overture service on getty
The mechanism in place now to switch between Overture and Dashboard
is through setting the default target: multi-user or graphical.
When Overture specifies that it is wanted by the getty service it
is sometimes started when we don't want it. This might fix the
issue users were having with updating and Overture starting
afterwards. More strangely is the fact that we were only able to
replicate the issue by reinstalling the package without this fix.